+++ src/cargo/core/source.rs
@@ -231,7 +231,7 @@ impl SourceId {
                 };
                 Box::new(PathSource::new(&path, self, config))
             }
-            Kind::Registry => Box::new(RegistrySource::new(self, config)),
+            Kind::Registry => Box::new(RegistrySource::remote(self, config)),
         }
     }
